backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
